Remove styled-components dependency
I have mixed feelings about this package. On one hand, I think it makes
really lovely view markup inside of render() blocks because each element
has an identifiable name. But the tradeoffs for that are that you're
forced to write CSS-style syntax, which is a dangerous mindset to get
into as React Native developers. React Native is not the web. I recently
discovered that there was a security vulnerability in the
styled-components package too, and so it made me reconsider the need to
have this as a production dependency. Are the trade-offs worth it still?
I don't think so. So I'm removing the package and going back to vanilla
styling.