Fix crash when using visbile prop to show/hide
As first reported at:
https://github.com/adammcarth/react-native-segmented-picker/issues/7

This fixes a crash which happened when using the `visible` prop to
control visibility. The library was trying to fade out the component
after it was being hidden due to a logic error in the previous
"isVisible()" function which returned true too early.

I have also implemented some updates to make the closing of the picker
smoother on iOS, regardless of whether you're using the `visible` prop
or the `ref.hide()` method directly.